LayoutTests/http/tests/security/http-0.9/default-port-plugin-blocked.html

<!DOCTYPE html>
<html>
<head>
<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.dumpChildFramesAsText();
    testRunner.waitUntilDone();
    if (window.internals)
        internals.registerDefaultPortForProtocol(8000, "http");
}

async function closeConnectionsThenAddFrame() {
    for (var i = 0; i < 6; i = i + 1) {
        try {
            await fetch('resources/close-connection.php', {mode: 'no-cors'});
            alert("successfully closed one connection");
        } catch (e) {
            alert("exception while closing connection " + e);
        }
    }
    var iframe = document.createElement('iframe');
    iframe.src = "resources/nph-load-plugin-fail.pl";
    iframe.onload = function() { alert("iframe loaded successfully"); if (window.testRunner) { testRunner.notifyDone() } }
    iframe.onerror = function() { alert("iframe load failed"); if (window.testRunner) { testRunner.notifyDone() } }
    document.body.appendChild(iframe);
}

</script>
</head>
<body onload="closeConnectionsThenAddFrame()">
<p id="description">This tests that loading of a plugin is blocked in a HTTP 0.9 response served over port 80. This test PASSED if you see content in the iframe below. Otherwise, it FAILED.</p>
</body>
</html>
